progress on the shortcut list, able to accept code
